Move db enums to be closer to the schema
