+++ app/suppliers/[handle]/new/upload-preview-form.tsx
@@ -7,7 +7,7 @@ import { Textarea } from '@/components/ui/textarea'
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select'
 import { Label } from '@/components/ui/label'
 import { InsertTileRaw, SupplierRaw, User } from '@/models/types'
-import { Location } from '@/models/constants'
+import { Location } from '@/db/constants'
 import { enumToPretty } from '@/utils/enum-to-pretty'
 import { SubmitButton } from '@/components/submit-button'
 import { useCreateTile } from '@/hooks/use-create-tile'
